deleting item_repository
