+++ lib/app/modules/shopping_list/domain/repositories/shopping_list_repository.dart
@@ -1,4 +1,5 @@
 import 'package:dartz/dartz.dart';
+import 'package:easy_market/app/modules/shopping_list/domain/entities/item.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/entities/shopping_list.dart';
 import 'package:easy_market/app/core/errors/errors.dart';
 
@@ -10,4 +11,10 @@ abstract class ShoppingListRepository {
       ShoppingList shoppingList);
   Future<Either<Failure, Unit>> updateShoppingList(ShoppingList shoppingList);
   Future<Either<Failure, Unit>> deleteShoppingList(ShoppingList shoppingList);
+  Future<Either<Failure, List<Item>>> getItemsFromList(String shoppingListId);
+  Either<Failure, Stream<List<Item>>> listenItemsFromList(
+      String shoppingListId);
+  Future<Either<Failure, Item>> addItemToList(Item item);
+  Future<Either<Failure, Unit>> updateItemInList(Item item);
+  Future<Either<Failure, Unit>> deleteItemFromList(Item item);
 }
