+++ lib/app/modules/shopping_list/domain/usecases/add_item_to_list.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/entities/item.dart';
 import 'package:easy_market/app/core/errors/errors.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/errors/errors.dart';
-import 'package:easy_market/app/modules/shopping_list/domain/repositories/item_repository.dart';
+import 'package:easy_market/app/modules/shopping_list/domain/repositories/shopping_list_repository.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 part 'add_item_to_list.g.dart';
 
@@ -12,7 +12,7 @@ abstract class AddItemToList {
 
 @Injectable(singleton: false)
 class AddItemToListImpl implements AddItemToList {
-  final ItemRepository repository;
+  final ShoppingListRepository repository;
   AddItemToListImpl(this.repository);
 
   @override
