+++ lib/app/modules/shopping_list/domain/usecases/delete_item_from_list.dart
@@ -1,9 +1,9 @@
 import 'package:dartz/dartz.dart';
+import 'package:easy_market/app/modules/shopping_list/domain/repositories/shopping_list_repository.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/entities/item.dart';
 import 'package:easy_market/app/core/errors/errors.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/errors/errors.dart';
-import 'package:easy_market/app/modules/shopping_list/domain/repositories/item_repository.dart';
 part 'delete_item_from_list.g.dart';
 
 abstract class DeleteItemFromList {
@@ -12,7 +12,7 @@ abstract class DeleteItemFromList {
 
 @Injectable(singleton: false)
 class DeleteItemFromListImpl implements DeleteItemFromList {
-  final ItemRepository repository;
+  final ShoppingListRepository repository;
   DeleteItemFromListImpl(this.repository);
 
   @override
