+++ lib/app/modules/shopping_list/domain/usecases/update_item_in_list.dart
@@ -1,9 +1,9 @@
 import 'package:dartz/dartz.dart';
+import 'package:easy_market/app/modules/shopping_list/domain/repositories/shopping_list_repository.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/entities/item.dart';
 import 'package:easy_market/app/core/errors/errors.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/errors/errors.dart';
-import 'package:easy_market/app/modules/shopping_list/domain/repositories/item_repository.dart';
 part 'update_item_in_list.g.dart';
 
 abstract class UpdateItemInList {
@@ -12,7 +12,7 @@ abstract class UpdateItemInList {
 
 @Injectable(singleton: false)
 class UpdateItemInListImpl implements UpdateItemInList {
-  final ItemRepository repository;
+  final ShoppingListRepository repository;
   UpdateItemInListImpl(this.repository);
 
   @override
