+++ lib/app/modules/shopping_list/infra/repositories/shopping_list_repository_impl.dart
@@ -1,3 +1,5 @@
+import 'package:easy_market/app/modules/shopping_list/domain/entities/item.dart';
+import 'package:easy_market/app/modules/shopping_list/infra/models/item_model.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:easy_market/app/core/errors/errors.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/entities/shopping_list.dart';
@@ -67,6 +69,60 @@ class ShoppingListRepositoryImpl implements ShoppingListRepository {
       final shoppingListToUpdate =
           ShoppingListModel.fromShoppingList(shoppingList);
       await datasource.updateShoppingList(shoppingListToUpdate);
+      return right(unit);
+    } catch (e) {
+      return left(ShoppingListFailure());
+    }
+  }
+
+  @override
+  Future<Either<Failure, List<Item>>> getItemsFromList(
+      String shoppingListId) async {
+    try {
+      var result = await datasource.getItemsFromList(shoppingListId);
+      return right(result);
+    } catch (e) {
+      return left(ShoppingListFailure());
+    }
+  }
+
+  @override
+  Either<Failure, Stream<List<Item>>> listenItemsFromList(
+      String shoppingListId) {
+    try {
+      var result = datasource.listenItemsFromList(shoppingListId);
+      return right(result);
+    } catch (e) {
+      return left(ShoppingListFailure());
+    }
+  }
+
+  @override
+  Future<Either<Failure, Item>> addItemToList(Item item) async {
+    try {
+      var itemToAdd = ItemModel.fromItem(item);
+      var result = await datasource.addItemToList(itemToAdd);
+      return right(result);
+    } catch (e) {
+      return left(ShoppingListFailure());
+    }
+  }
+
+  @override
+  Future<Either<Failure, Unit>> updateItemInList(Item item) async {
+    try {
+      var itemToUpdate = ItemModel.fromItem(item);
+      await datasource.updateItemInList(itemToUpdate);
+      return right(unit);
+    } catch (e) {
+      return left(ShoppingListFailure());
+    }
+  }
+
+  @override
+  Future<Either<Failure, Unit>> deleteItemFromList(Item item) async {
+    try {
+      await datasource.deleteItemFromList(item.id);
       return right(unit);
     } catch (e) {
       return left(ShoppingListFailure());
