+++ lib/app/modules/shopping_list/shopping_list.dart
@@ -6,9 +6,7 @@ export 'domain/usecases/add_item_to_list.dart';
 export 'domain/usecases/create_shopping_list.dart';
 export 'domain/usecases/delete_item_from_list.dart';
 export 'domain/usecases/delete_shopping_list.dart';
-export 'domain/usecases/get_items_from_list.dart';
 export 'domain/usecases/get_shopping_lists.dart';
-export 'domain/usecases/listen_items_from_list.dart';
 export 'domain/usecases/listen_shopping_lists.dart';
 export 'domain/usecases/update_item_in_list.dart';
 export 'domain/usecases/update_shopping_list.dart';
