+++ lib/app/modules/shopping_list/shopping_list_module.dart
@@ -12,7 +12,6 @@ import 'package:easy_market/app/modules/shopping_list/domain/usecases/listen_sho
 import 'package:easy_market/app/modules/shopping_list/domain/usecases/update_item_in_list.dart';
 import 'package:easy_market/app/modules/shopping_list/domain/usecases/update_shopping_list.dart';
 import 'package:easy_market/app/modules/shopping_list/external/datasources/firebase/firebase_shopping_list_datasource.dart';
-import 'package:easy_market/app/modules/shopping_list/infra/repositories/item_repository_impl.dart';
 import 'package:easy_market/app/modules/shopping_list/infra/repositories/shopping_list_repository_impl.dart';
 import 'package:easy_market/app/modules/shopping_list/presenter/bloc/shopping_list_bloc.dart';
 import 'package:easy_market/app/modules/shopping_list/presenter/pages/shopping_list_details_page.dart';
@@ -32,7 +31,6 @@ class ShoppingListModule extends Module {
     $ListenShoppingListsImpl,
     $UpdateItemInListImpl,
     $UpdateShoppingListImpl,
-    $ItemRepositoryImpl,
     $ShoppingListRepositoryImpl,
     BlocBind.singleton(
       (i) => ShoppingListBloc(i<ListenShoppingLists>()),
